do the validation in PackageBuilder instead
