+++ src/NuGet.Core/NuGet.Packaging/Rules/RuleSet.cs
@@ -30,7 +30,6 @@ namespace NuGet.Packaging.Rules
                 new UpholdBuildConventionRule(),
                 new ReferencesInNuspecMatchRefAssetsRule(),
                 new InvalidUndottedFrameworkRule(),
-                new TargetFrameworkMissingPlatformVersionRule(),
                 new IconUrlDeprecationWarning(AnalysisResources.IconUrlDeprecationWarning),
             }
         );
